[Librarian] Regenerated @ 125dc871b4170bce0faca169f8a588f6ff356a88 0c222649d98edbac6858ea253f9b77fc22386e47

+++ src/main/java/com/twilio/rest/api/v2010/AccountReader.java
@@ -31,7 +31,7 @@ public class AccountReader extends Reader<Account> {
 
     private String friendlyName;
     private Account.Status status;
-    private Integer pageSize;
+    private Long pageSize;
 
     public AccountReader() {}
 
@@ -45,7 +45,7 @@ public class AccountReader extends Reader<Account> {
         return this;
     }
 
-    public AccountReader setPageSize(final Integer pageSize) {
+    public AccountReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
