+++ src/main/java/com/twilio/rest/api/v2010/account/NotificationReader.java
@@ -36,7 +36,7 @@ public class NotificationReader extends Reader<Notification> {
     private LocalDate messageDate;
     private LocalDate messageDateBefore;
     private LocalDate messageDateAfter;
-    private Integer pageSize;
+    private Long pageSize;
 
     public NotificationReader() {}
 
@@ -68,7 +68,7 @@ public class NotificationReader extends Reader<Notification> {
         return this;
     }
 
-    public NotificationReader setPageSize(final Integer pageSize) {
+    public NotificationReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
