+++ src/main/java/com/twilio/rest/chat/v1/CredentialReader.java
@@ -29,11 +29,11 @@ import com.twilio.rest.Domains;
 
 public class CredentialReader extends Reader<Credential> {
 
-    private Integer pageSize;
+    private Long pageSize;
 
     public CredentialReader() {}
 
-    public CredentialReader setPageSize(final Integer pageSize) {
+    public CredentialReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
