+++ src/main/java/com/twilio/rest/conversations/v1/UserReader.java
@@ -29,11 +29,11 @@ import com.twilio.rest.Domains;
 
 public class UserReader extends Reader<User> {
 
-    private Integer pageSize;
+    private Long pageSize;
 
     public UserReader() {}
 
-    public UserReader setPageSize(final Integer pageSize) {
+    public UserReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
