+++ src/main/java/com/twilio/rest/conversations/v1/conversation/MessageReader.java
@@ -31,7 +31,7 @@ public class MessageReader extends Reader<Message> {
 
     private String pathConversationSid;
     private Message.OrderType order;
-    private Integer pageSize;
+    private Long pageSize;
 
     public MessageReader(final String pathConversationSid) {
         this.pathConversationSid = pathConversationSid;
@@ -42,7 +42,7 @@ public class MessageReader extends Reader<Message> {
         return this;
     }
 
-    public MessageReader setPageSize(final Integer pageSize) {
+    public MessageReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
