+++ src/main/java/com/twilio/rest/conversations/v1/conversation/ParticipantReader.java
@@ -30,13 +30,13 @@ import com.twilio.rest.Domains;
 public class ParticipantReader extends Reader<Participant> {
 
     private String pathConversationSid;
-    private Integer pageSize;
+    private Long pageSize;
 
     public ParticipantReader(final String pathConversationSid) {
         this.pathConversationSid = pathConversationSid;
     }
 
-    public ParticipantReader setPageSize(final Integer pageSize) {
+    public ParticipantReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
