+++ src/main/java/com/twilio/rest/conversations/v1/conversation/message/DeliveryReceiptReader.java
@@ -31,7 +31,7 @@ public class DeliveryReceiptReader extends Reader<DeliveryReceipt> {
 
     private String pathConversationSid;
     private String pathMessageSid;
-    private Integer pageSize;
+    private Long pageSize;
 
     public DeliveryReceiptReader(
         final String pathConversationSid,
@@ -41,7 +41,7 @@ public class DeliveryReceiptReader extends Reader<DeliveryReceipt> {
         this.pathMessageSid = pathMessageSid;
     }
 
-    public DeliveryReceiptReader setPageSize(final Integer pageSize) {
+    public DeliveryReceiptReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
