+++ src/main/java/com/twilio/rest/conversations/v1/service/RoleReader.java
@@ -30,13 +30,13 @@ import com.twilio.rest.Domains;
 public class RoleReader extends Reader<Role> {
 
     private String pathChatServiceSid;
-    private Integer pageSize;
+    private Long pageSize;
 
     public RoleReader(final String pathChatServiceSid) {
         this.pathChatServiceSid = pathChatServiceSid;
     }
 
-    public RoleReader setPageSize(final Integer pageSize) {
+    public RoleReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
