+++ src/main/java/com/twilio/rest/insights/v1/call/EventReader.java
@@ -31,7 +31,7 @@ public class EventReader extends Reader<Event> {
 
     private String pathCallSid;
     private Event.TwilioEdge edge;
-    private Integer pageSize;
+    private Long pageSize;
 
     public EventReader(final String pathCallSid) {
         this.pathCallSid = pathCallSid;
@@ -42,7 +42,7 @@ public class EventReader extends Reader<Event> {
         return this;
     }
 
-    public EventReader setPageSize(final Integer pageSize) {
+    public EventReader setPageSize(final Long pageSize) {
         this.pageSize = pageSize;
         return this;
     }
